Refs #16784. Small refactor.
